+++ TIFoundationUtils/AsyncOperation/Sources/AsyncOperation+Observe.swift
@@ -22,46 +22,45 @@
 
 import Foundation
 
-private final class ClosureObserverOperation<Output, Failure: Error>: AsyncOperation<Output, Failure> {
-    private var dependencyObservation: NSKeyValueObservation?
+private final class ClosureObserverOperation<Output, Failure: Error>: DependendAsyncOperation<Output, Failure> {
+    public typealias OnResultClosure = (Result<Output, Failure>) -> Void
 
     public init(dependency: AsyncOperation<Output, Failure>,
-                onSuccess: ((Output) -> Void)? = nil,
-                onFailure: ((Failure) -> Void)? = nil,
+                onResult: OnResultClosure? = nil,
                 callbackQueue: DispatchQueue = .main) {
 
-        super.init()
-
-        cancelOnCancellation(of: dependency)
-
-        dependencyObservation = dependency.subscribe { [weak self] result in
+        super.init(dependency: dependency) { result in
             callbackQueue.async {
-                onSuccess?(result)
+                onResult?(result)
             }
-
-            self?.handle(result: result)
-        } onFailure: { [weak self] error in
-            callbackQueue.async {
-                onFailure?(error)
-            }
-
-            self?.handle(error: error)
+            return result
         }
-
-        addDependency(dependency) // keeps strong reference to dependency as well
-
-        state = .isReady
     }
 }
 
 public extension AsyncOperation {
+    func observe(onResult: ((Result<Output, Failure>) -> Void)? = nil,
+                 callbackQueue: DispatchQueue = .main) -> AsyncOperation<Output, Failure> {
+
+        ClosureObserverOperation(dependency: self,
+                                 onResult: onResult,
+                                 callbackQueue: callbackQueue)
+    }
+
     func observe(onSuccess: ((Output) -> Void)? = nil,
                  onFailure: ((Failure) -> Void)? = nil,
                  callbackQueue: DispatchQueue = .main) -> AsyncOperation<Output, Failure> {
 
-        ClosureObserverOperation(dependency: self,
-                                 onSuccess: onSuccess,
-                                 onFailure: onFailure,
-                                 callbackQueue: callbackQueue)
+        let onResult: ClosureObserverOperation<Output, Failure>.OnResultClosure = {
+            switch $0 {
+            case let .success(output):
+                onSuccess?(output)
+
+            case let .failure(error):
+                onFailure?(error)
+            }
+        }
+
+        return observe(onResult: onResult, callbackQueue: callbackQueue)
     }
 }